* Add support to boot loopback mounted ISO images on the recovery partition.   - They must be titled 'ubuntu.iso' to function because some parts of the     code can't dynamically detect the filename. * Run oem-config early and late commands from a shell script rather than a   long chained list of && commands. * Remove the functionality of setting the pool command. * Add a new black_tree function that copies files off of a blacklist.

+++ late/scripts/pool.sh
@@ -63,9 +63,16 @@ if [ ! -f /etc/apt/sources.list.d/dell.list ]; then
         mv /etc/apt/sources.list.d/* /etc/apt/sources.list.d.old
     fi
     #Produce a dynamic list
-    cd /cdrom/debs
-    apt-ftparchive packages ../../cdrom/debs | sed "s/^Filename:\ ..\//Filename:\ .\//" > /Packages
-    echo "deb file:/ /" > /etc/apt/sources.list.d/dell.list
+    for dir in /cdrom/debs /isodevice/debs;
+    do
+        if [ -d "$dir" ]; then
+            cd $dir
+            apt-ftparchive packages ../../$dir | sed "s/^Filename:\ ..\//Filename:\ .\//" >> /Packages
+        fi
+    done
+    if [ -f /Packages ]; then
+        echo "deb file:/ /" > /etc/apt/sources.list.d/dell.list
+    fi
 
     #add the static list to our file
     apt-cdrom -m add
